add data download on building and collection pages (#31)
This add csv download button for buildings pages that gives all the indicators for that page (not including chart data) transposed for readability and includes the full indicator names and descriptions.

Also adds download for collection pages for all the buildings data (filtered version will be separate). This includes two header rows for api names and full indicator names (no descriptions). Also changes the lender/landlord slug variables into links to those pages on the dashboard.

In a separate PR we'll also need to add a button for downloading all buildings. Waiting to see if this will be on a page where we are also showing that table (so the data is already downloaded on page load) or not since we'll need to figure out how to conditionally download the data to the client before preparing the csv download.