add Algolia search bars for address and landlord (#35)
This changes our method for the address and landlord typeahead/autocomplete search bars to use Algolia. This is service we already use on WOW for search by landlord name. We have some separate scripts (we can consider adding them here later) to generate the data, then upload it to algolia, which then allows us to query and get back search results. We copy the same setup for the dropdown menu with aria announcements from WOW, but adapt things slightly to work with the latest version of the algolia packages (we are on v3 in wow, and here using v7). I've made the styles match our component library dropdown.

For the address data we use the PAD database of addresses to get all the valid street addresses for each BBL, using the code from the Geosearch API so this should work just like that one but for only our buildings. (for details see https://github.com/JustFixNYC/requests/blob/main/signature-dashboard/address-data-prep-guide.md)

We do the same thing with all the landlord names.

A couple remaining issues to fix in a later PR:

add the algolia logo to the dropdowns (same as in wow)
fix the aria setup for screen reader - user input changes and search results announcements interrupt each other