Add link to map on building pages (#40)
Adds bbl search params to map page and adds links from building pages. Map now has optional props for setting initial selected point (with side pane open). Once the map loads it then clears the url query params (considered just replacing the url to prevent cluttering the browser history, but since they are not meant for sharing decide to just remove it).

Also this changes so that clicks on the map that aren't on a point no longer clears the selected address, since it's easy to do this inadvertently when panning around.