Add maps to collection pages (#47)
Adds a map of buildings to collection pages (only landlord and lender, not entire signature portfolio). Puts it above the buildings table, with a section header including the lender/landlord name like we're change the building table one. Adds an option to disable regular scroll zoom on the map since this one's on a page with scroll (you can use cmd+scroll to zoom or pinch). Also fixes the width of the sidepane for selected buildings. And I split out the styles creating a separate file for the map component separate from the map/collection page styles.