separate aggregate indicators for collection pages (#59)
there are some aggregate indicators for collection pages that use the same name as the building level ones (eg. units_res), but in some cases we want to have more explicit indicator descriptions that make it clear it covers all the signature-financed buildings in that portfolio (and not all the landlords properties, for example). So we are renaming all of these variables with a _agg suffix to allow for this.

unrelated, but also add unhp as dataset for indicators they provide us (these last_updated dates need to be manually updated in the database)

Also unrelated, I updated the OCA data note on collection page charts to state when the collection includes buildings with < 11 units. (before it was mistakenly using the total collection count of units - see #58)