turn off dropdown option filtering (#67)
For the Good Cause Eviction screener we got user feedback that the address search was acting buggy. It turns out that the reac-select component we use for the dropdown does its own filtering of the options based on input (JustFixNYC/gce-screener#13), but when using Geosearch or Algolia those are handling the filtering of options and they produce better results. For example, putting something like 8th ave would cause react-select to remove results without the th whereas the API results include them.