Add loan status for all buildings (#70)
This PR adds the new loan status feature to the dashboard that we worked on with UNHP. Ana has completed a review of all the properties in the portfolio and added status information to the building data, and we have added new design elements to the building pages to display that and updated the SQL queries to work with buildings leaving the program and other changes.

Adds a section for loan status and history to all building pages
Adds a current status column to all building tables
Excludes properties that leave the program from aggregate stats (eg. portfolio pages)
Updates API documentation for new fields
Adds a note to "about" section mentioning the third loan pool under Blackrock.
Companion PRs:

update Signature dashboard backend data/api for loan status who-owns-what#1035
Update WOW for Signature loan status nycdb-k8s-loader#213