make loan status last updated date dynamic (#71)
Before this was hardcoded, but instead we can use the same "unhp" "dataset" from the dataset tracker system we have for other datasets. This one we just manually added to the database, and update the date manually (eg. via postico) when we get a new set of data from UNHP.